Adjust to the core. Inserting @User entity and Base interfaces.

+++ src/main/java/net/raza/core/controllers/ProductController.java
@@ -28,13 +28,13 @@ public class ProductController {
     }
 
     @RequestMapping("product/{id}")
-    public String showProduct(@PathVariable Integer id, Model model){
+    public String showProduct(@PathVariable Long id, Model model){
         model.addAttribute("product", productService.getProductById(id));
         return "productshow";
     }
 
     @RequestMapping("product/edit/{id}")
-    public String edit(@PathVariable Integer id, Model model){
+    public String edit(@PathVariable Long id, Model model){
         model.addAttribute("product", productService.getProductById(id));
         return "productform";
     }
